requirements.txt
fastapi
uvicorn
google-generativeai
pandas
openpyxl
streamlit
python-multipart